refactor: give Outputter its own folder
